v2025.06.25.2 Add check for getting version

+++ pycharm_install_script.sh
@@ -11,20 +11,25 @@ sudo rm -Rf /home/$USER/.local/share/applications/jetbrains-pycharm*
 # Getting PyCharm version from official site
 echo -e "Getting PyCharm version from official site...\n"
 PYCHARM_VERSION=$(curl -s "https://data.services.jetbrains.com/products/releases?code=PCC&latest=true" | jq -r '.["PCC"][0].version')
-echo -e "PyCharm version is $PYCHARM_VERSION\n"
-echo -e "Downloading...\n"
-wget "https://download.jetbrains.com/python/pycharm-community-$PYCHARM_VERSION.tar.gz" -O PyCharm.tar.gz
-echo -e "\nInstalling...\n"
-sudo tar -xzf PyCharm.tar.gz -C /opt
-sudo ln -sf /opt/pycharm-community-$PYCHARM_VERSION/bin/pycharm.sh /usr/bin/pycharm
-sudo rm -Rf PyCharm.tar.gz
-echo -e "PyCharm was installed!\n"
+if [ -n "$PYCHARM_VERSION" ] && [ "$PYCHARM_VERSION" != "null" ]; then
+  echo -e "PyCharm version is $PYCHARM_VERSION\n"
+  echo -e "Downloading...\n"
+  wget "https://download.jetbrains.com/python/pycharm-community-$PYCHARM_VERSION.tar.gz" -O PyCharm.tar.gz
+  echo -e "\nInstalling...\n"
+  sudo tar -xzf PyCharm.tar.gz -C /opt
+  sudo ln -sf /opt/pycharm-community-$PYCHARM_VERSION/bin/pycharm.sh /usr/bin/pycharm
+  sudo rm -Rf PyCharm.tar.gz
+  echo -e "PyCharm was installed!\n"
 
-if [ "$1" = "boot" ]
-then
-  echo -e "Initial launch of PyCharm\n"
-  echo -e "In welcome settings or Tools you can create desktop entry"
-  pycharm
+  if [ "$1" = "boot" ]
+  then
+    echo -e "Initial launch of PyCharm\n"
+    echo -e "In welcome settings or Tools you can create desktop entry"
+    pycharm
+  else
+    echo -e "Type pycharm in terminal and in welcome settings or Tools you can create desktop entry"
+  fi
 else
-  echo -e "Type pycharm in terminal and in welcome settings or Tools you can create desktop entry"
+  echo -e "Can't get PyCharm version from official site...\n"
+  exit 1
 fi
